feat: [US-002] - Fix all clippy warnings
- Use std::array::from_fn for AtomicU64 array initialization (metrics.rs)
- Replace .get().is_none() with !contains_key() (acc_collector.rs)
- Combine duplicate if branches with || operator (stitcher.rs)
- Remove unnecessary u64 casts (handlers.rs)
- Prefix unused struct fields with underscore
- Add #[allow(dead_code)] for public API methods not yet used
- Add #[allow(clippy::too_many_arguments)] for internal function
- Run cargo clippy --fix for auto-fixable warnings

+++ src/services/tracker/handlers.rs
@@ -38,7 +38,7 @@ impl Tracker {
             let StitchMatch { mut person, time_ms, distance_cm } = stitch;
             self.metrics.record_stitch_matched();
             self.metrics.record_stitch_distance(distance_cm as u64);
-            self.metrics.record_stitch_time(time_ms as u64);
+            self.metrics.record_stitch_time(time_ms);
 
             // Stitch found! Transfer state from old track to new track
             let old_track_id = person.track_id;
@@ -66,7 +66,7 @@ impl Tracker {
                     auth: person.authorized,
                     dwell_ms: person.accumulated_dwell_ms,
                     stitch_dist_cm: Some(distance_cm as u64),
-                    stitch_time_ms: Some(time_ms as u64),
+                    stitch_time_ms: Some(time_ms),
                     parent_jid: None,
                 });
             }
